feat: page 와 관련된 state를 관리하는 커스텀 훅 구현
