+++ src/hooks/useProductList.ts
@@ -1,38 +1,21 @@
 import { useEffect, useState } from 'react';
 
 import useProductDropdown from './useProductDropdown';
+import useProductPage from './useProductPage';
+import useToast from './useToast';
 
 import { fetchProductList } from '@/api/product';
+import { PAGE_SIZE } from '@/constants/config';
 import { Product } from '@/types/product';
 import CustomError from '@/utils/error';
-import { PAGE_SIZE } from '@/constants/config';
-import useToast from './useToast';
-import useError from './useError';
 
 const useProductList = () => {
   const [productList, setProductList] = useState<Product[]>([]);
-  const [page, setPage] = useState(0);
-  const [isLastPage, setIsLastPage] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
-  const { errorState, handleError, resetError } = useError();
 
   const toast = useToast();
-  const { category, order, handleChangeCategory, handleChangeSort } = useProductDropdown({
-    resetPage: () => setPage(0),
-  });
-
-  const fetchNextPage = () => {
-    const nextPageUnit = page === 0 ? PAGE_SIZE.firstPageUnit : PAGE_SIZE.nextPageUnit;
-    if (isLastPage) return;
-
-    setPage((prevPage) => prevPage + nextPageUnit);
-  };
-
-  useEffect(() => {
-    window.addEventListener('online', () => {
-      resetError();
-    });
-  }, []);
+  const { page, fetchNextPage, determineLastPage, resetPage } = useProductPage();
+  const { category, order, handleChangeCategory, handleChangeSort } = useProductDropdown(resetPage);
 
   useEffect(() => {
     const getProductList = async () => {
@@ -45,16 +28,11 @@ const useProductList = () => {
           size: limit,
           sortOptions: order.value,
         });
-        if (last) {
-          setIsLastPage(true);
-        } else {
-          setIsLastPage(false);
-        }
+
+        determineLastPage(last);
         page === 0 ? setProductList(content) : setProductList((prev) => [...prev, ...content]);
-        resetError();
       } catch (error) {
         if (error instanceof CustomError) {
-          handleError({ isError: true, name: error.name, errorMessage: error.message });
           toast.error(error.message);
         }
       } finally {
@@ -74,9 +52,6 @@ const useProductList = () => {
     handleChangeSort,
     category,
     order,
-    errorState,
-    handleError,
-    resetError,
   };
 };
